Adds the TraitSelectDialog class and svelte
This class allows you to generate customized trait selection. It requires you provide a list of of traitGroups (For example, simple, martial and rare weapons), along with the label for that group.

If a traitGroup includes no traits, it will assume you are attempting to create a "custom" grouping, which allows the user to enter a comma seperated list of elements.

It also has a number of option parameters:

* options.selected - Defines the list of preselected elements.
* options.disabled - Defines a list of traits which selection status is disabled.
* options.selectionLimit - Requires the user to selected elements + an additional amounts from being added. A negative amount would require the user to remove selections.

The class resolves it's `promise` property when the user has submited their choices